+++ configs/experiment/gpt2_baseline.yaml
@@ -1,0 +1,6 @@
+name: gpt2_baseline
+description: GPT-2 baseline on cleaned MIDI corpus.
+model_config: configs/model/gpt2/nano.yaml
+tokenizer_config: configs/data/tokenizer_config.yaml
+train_config: configs/train/base.yaml
+eval_config: configs/eval/music_metrics.yaml
