+++ configs/model/xlstm/basic.yaml
