+++ configs/model/xlstm/ithinkadvanced.yaml
