+++ configs/model/xlstm/ithinkthisistherightone.yaml
