--- configs/preprocessing/tokenizer_config.yaml
@@ -1,425 +1,0 @@
-tokenizer: REMI
-tokenizer_config:
-  pitch_range: !!python/tuple
-  - 21
-  - 109
-  beat_res:
-    ? !!python/tuple
-    - 0
-    - 4
-    : 8
-    ? !!python/tuple
-    - 4
-    - 12
-    : 4
-  num_velocities: 32
-  special_tokens:
-  - PAD
-  - BOS
-  - EOS
-  - MASK
-  encode_ids_split: bar
-  use_velocities: true
-  use_note_duration_programs:
-  - -1
-  - 0
-  - 1
-  - 2
-  - 3
-  - 4
-  - 5
-  - 6
-  - 7
-  - 8
-  - 9
-  - 10
-  - 11
-  - 12
-  - 13
-  - 14
-  - 15
-  - 16
-  - 17
-  - 18
-  - 19
-  - 20
-  - 21
-  - 22
-  - 23
-  - 24
-  - 25
-  - 26
-  - 27
-  - 28
-  - 29
-  - 30
-  - 31
-  - 32
-  - 33
-  - 34
-  - 35
-  - 36
-  - 37
-  - 38
-  - 39
-  - 40
-  - 41
-  - 42
-  - 43
-  - 44
-  - 45
-  - 46
-  - 47
-  - 48
-  - 49
-  - 50
-  - 51
-  - 52
-  - 53
-  - 54
-  - 55
-  - 56
-  - 57
-  - 58
-  - 59
-  - 60
-  - 61
-  - 62
-  - 63
-  - 64
-  - 65
-  - 66
-  - 67
-  - 68
-  - 69
-  - 70
-  - 71
-  - 72
-  - 73
-  - 74
-  - 75
-  - 76
-  - 77
-  - 78
-  - 79
-  - 80
-  - 81
-  - 82
-  - 83
-  - 84
-  - 85
-  - 86
-  - 87
-  - 88
-  - 89
-  - 90
-  - 91
-  - 92
-  - 93
-  - 94
-  - 95
-  - 96
-  - 97
-  - 98
-  - 99
-  - 100
-  - 101
-  - 102
-  - 103
-  - 104
-  - 105
-  - 106
-  - 107
-  - 108
-  - 109
-  - 110
-  - 111
-  - 112
-  - 113
-  - 114
-  - 115
-  - 116
-  - 117
-  - 118
-  - 119
-  - 120
-  - 121
-  - 122
-  - 123
-  - 124
-  - 125
-  - 126
-  - 127
-  use_chords: false
-  use_rests: false
-  use_tempos: false
-  use_time_signatures: false
-  use_sustain_pedals: false
-  use_pitch_bends: false
-  use_programs: false
-  use_pitch_intervals: false
-  use_pitchdrum_tokens: true
-  default_note_duration: 0.5
-  beat_res_rest:
-    ? !!python/tuple
-    - 0
-    - 1
-    : 8
-    ? !!python/tuple
-    - 1
-    - 2
-    : 4
-    ? !!python/tuple
-    - 2
-    - 12
-    : 2
-  chord_maps:
-    maj: !!python/tuple
-    - 0
-    - 4
-    - 7
-    min: !!python/tuple
-    - 0
-    - 3
-    - 7
-    aug: !!python/tuple
-    - 0
-    - 4
-    - 8
-    dim: !!python/tuple
-    - 0
-    - 3
-    - 6
-    7maj: !!python/tuple
-    - 0
-    - 4
-    - 7
-    - 11
-    7min: !!python/tuple
-    - 0
-    - 3
-    - 7
-    - 10
-    7dom: !!python/tuple
-    - 0
-    - 4
-    - 7
-    - 10
-    7halfdim: !!python/tuple
-    - 0
-    - 3
-    - 6
-    - 10
-    7aug: !!python/tuple
-    - 0
-    - 4
-    - 8
-    - 11
-    7dim: !!python/tuple
-    - 0
-    - 3
-    - 6
-    - 9
-    9maj: !!python/tuple
-    - 0
-    - 4
-    - 7
-    - 10
-    - 14
-    9min: !!python/tuple
-    - 0
-    - 4
-    - 7
-    - 10
-    - 13
-    sus2: !!python/tuple
-    - 0
-    - 2
-    - 7
-    sus4: !!python/tuple
-    - 0
-    - 5
-    - 7
-  chord_tokens_with_root_note: false
-  chord_unknown: null
-  num_tempos: 32
-  tempo_range: !!python/tuple
-  - 40
-  - 250
-  log_tempos: false
-  delete_equal_successive_tempo_changes: false
-  remove_duplicated_notes: false
-  time_signature_range:
-    8:
-    - 3
-    - 12
-    - 6
-    4:
-    - 5
-    - 6
-    - 3
-    - 2
-    - 1
-    - 4
-  sustain_pedal_duration: false
-  pitch_bend_range: !!python/tuple
-  - -8192
-  - 8191
-  - 32
-  delete_equal_successive_time_sig_changes: false
-  programs:
-  - -1
-  - 0
-  - 1
-  - 2
-  - 3
-  - 4
-  - 5
-  - 6
-  - 7
-  - 8
-  - 9
-  - 10
-  - 11
-  - 12
-  - 13
-  - 14
-  - 15
-  - 16
-  - 17
-  - 18
-  - 19
-  - 20
-  - 21
-  - 22
-  - 23
-  - 24
-  - 25
-  - 26
-  - 27
-  - 28
-  - 29
-  - 30
-  - 31
-  - 32
-  - 33
-  - 34
-  - 35
-  - 36
-  - 37
-  - 38
-  - 39
-  - 40
-  - 41
-  - 42
-  - 43
-  - 44
-  - 45
-  - 46
-  - 47
-  - 48
-  - 49
-  - 50
-  - 51
-  - 52
-  - 53
-  - 54
-  - 55
-  - 56
-  - 57
-  - 58
-  - 59
-  - 60
-  - 61
-  - 62
-  - 63
-  - 64
-  - 65
-  - 66
-  - 67
-  - 68
-  - 69
-  - 70
-  - 71
-  - 72
-  - 73
-  - 74
-  - 75
-  - 76
-  - 77
-  - 78
-  - 79
-  - 80
-  - 81
-  - 82
-  - 83
-  - 84
-  - 85
-  - 86
-  - 87
-  - 88
-  - 89
-  - 90
-  - 91
-  - 92
-  - 93
-  - 94
-  - 95
-  - 96
-  - 97
-  - 98
-  - 99
-  - 100
-  - 101
-  - 102
-  - 103
-  - 104
-  - 105
-  - 106
-  - 107
-  - 108
-  - 109
-  - 110
-  - 111
-  - 112
-  - 113
-  - 114
-  - 115
-  - 116
-  - 117
-  - 118
-  - 119
-  - 120
-  - 121
-  - 122
-  - 123
-  - 124
-  - 125
-  - 126
-  - 127
-  one_token_stream_for_programs: true
-  program_changes: false
-  max_pitch_interval: 16
-  pitch_intervals_max_time_dist: 1.0
-  drums_pitch_range: !!python/tuple
-  - 27
-  - 88
-  ac_polyphony_track: false
-  ac_polyphony_bar: false
-  ac_polyphony_min: 1
-  ac_polyphony_max: 6
-  ac_pitch_class_bar: false
-  ac_note_density_track: false
-  ac_note_density_track_min: 0
-  ac_note_density_track_max: 18
-  ac_note_density_bar: false
-  ac_note_density_bar_max: 18
-  ac_note_duration_bar: false
-  ac_note_duration_track: false
-  ac_repetition_track: false
-  ac_repetition_track_num_bins: 10
-  ac_repetition_track_num_consec_bars: 4
-additional_params:
-  max_bar_embedding: 512
-  use_bar_end_tokens: True
-  add_trailing_bars: false
