+++ configs/train/base.yaml
@@ -1,0 +1,14 @@
+seed: 42
+output_dir: experiments/latest/checkpoints
+train:
+  per_device_train_batch_size: 2
+  per_device_eval_batch_size: 2
+  learning_rate: 3e-4
+  num_train_epochs: 1
+  eval_strategy: steps
+  eval_steps: 100
+  save_steps: 100
+  logging_steps: 20
+  report_to: ["none"]
+data:
+  block_size: 1024
