+++ docs/metric_definitions.md
@@ -1,0 +1,6 @@
+# Metric Definitions
+
+- `precision`, `recall`, `f1`: onset event matching within tolerance.
+- `onset_mae_sec`: absolute timing error of matched note onsets.
+- `dur_mae_sec`: absolute error in median note duration per onset.
+- `chroma_dtw`: harmonic similarity score using DTW on chroma sequences.
