+++ experiments/.templates/README.md
@@ -1,0 +1,7 @@
+Create one directory per run, e.g. `experiments/2026-02-19-gpt2-vs-xlstm/`, with:
+- `config_snapshot/`
+- `logs/`
+- `checkpoints/`
+- `samples/`
+- `metrics.json`
+- `notes.md`
