+++ pyproject.toml
@@ -1,18 +1,7 @@
 [project]
 name = "duet-of-models"
 version = "0.1.0"
-description = "Utilities for cleaning and tokenizing MIDI data for duet model experiments."
+description = "Workspace metadata for duet-of-models. Use envs/* for runnable dependency sets."
 readme = "README.md"
 requires-python = ">=3.10"
-dependencies = [
-    "miditoolkit>=0.1.16",
-    "miditok>=3.0.0",
-    "PyYAML>=6.0",
-    "tqdm>=4.66",
-    "numpy>=2.2.6",
-    "pandas>=2.3.3",
-    "pretty-midi>=0.2.11",
-    "dtw-python>=1.7.2",
-    "scikit-learn>=1.7.2",
-    "jupyter>=1.1.1",
-]
+dependencies = []
