+++ test.py
@@ -1,0 +1,1 @@
+from tests.test_metrics_smoke import *  # noqa: F401,F403
